pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = [
    {name = "Kevin Gonzalez",email = "rkevingonzalez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pylint (>=4.0.3,<5.0.0)",
    "pydocstyle (>=6.3.0,<7.0.0)",
    "bandit (>=1.9.2,<2.0.0)"
]
